avances en producto

+++ lib/src/presentation/pages/subcategory/list/SubCategoryListItem.dart
@@ -2,6 +2,7 @@ import 'package:ecommerce_prueba/src/domain/models/Category.dart';
 import 'package:ecommerce_prueba/src/domain/models/SubCategory.dart';
 import 'package:ecommerce_prueba/src/presentation/pages/subcategory/list/bloc/SubCategoryListBloc.dart';
 import 'package:ecommerce_prueba/src/presentation/pages/subcategory/list/bloc/SubCategoryListEvent.dart';
+import 'package:ecommerce_prueba/src/presentation/utils/SelectConfirmDialog.dart';
 import 'package:flutter/material.dart';
 
 class SubCategoryListItem extends StatelessWidget {
@@ -13,7 +14,6 @@ class SubCategoryListItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    print('CATEGORIA listitem ${category.toJson()}');
     return Card(
       margin: EdgeInsets.symmetric(vertical: 3),
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
@@ -23,7 +23,7 @@ class SubCategoryListItem extends StatelessWidget {
           style: TextStyle(fontSize: 16, fontWeight: FontWeight.w500),
         ),
         subtitle: Text(
-          category.nombre!,
+          category.nombre,
           style: TextStyle(fontSize: 13, color: Colors.black),
         ),
         trailing: Row(
@@ -41,7 +41,11 @@ class SubCategoryListItem extends StatelessWidget {
             ),
             IconButton(
               onPressed: () {
-                bloc?.add(DeleteSubCategoryListEvent(id: subCategory.id));
+                selectConfirmDialog(
+                  context,
+                  () =>
+                      bloc?.add(DeleteSubCategoryListEvent(id: subCategory.id)),
+                );
               },
               icon: Icon(Icons.delete, color: Colors.red),
             ),
